refactor: choice wizard node persistence 분리
