+++ src/pages/workflow-editor/model/index.ts
@@ -1,3 +1,4 @@
 export * from "./choiceSelectionPipeline";
 export * from "./choiceWizardLogger";
+export * from "./choiceWizardNodePersistence";
 export * from "./useChoiceWizardController";
